requirements.txt
# System Design Interview Companion Dependencies
langchain
langchain-openai
python-dotenv
